/main

+++ src/service/PatientService.java
@@ -9,7 +9,7 @@ import java.util.Map;
 public interface PatientService extends GenericService<Patient>{
     String addPatientsToHospital(Long id, List<Patient> patients);
     Patient getPatientById(Long hospitalId,Long id);
-    Map<Integer, Patient> getPatientByAge(Long hospitalId, Integer age);
+    Map<Long, Patient> getPatientByAge(Long hospitalId, Integer age);
     List<Patient> sortPatientsByAge(Long hospitalId,String ascOrDesc);
 }
